[repairability] new test classes and patterns file reformate
