+++ src/main/resources/repairability/Nopol/if_upd_d.xml
@@ -1,17 +1,15 @@
 <pattern name="if_condition_upd">
 
     <entity id="1" type="If"/>
-
     <entity id="2" type="*" role="condition">
         <parent parentId="1" distance="1"/>
     </entity>
-
     <entity id="4" type="*">
         <parent parentId="2" distance="1000000"/>
     </entity>
     <action entityId="4" type="*"/>
 
-    <entity id="3" type="Block" role="Then">
+    <entity id="3" type="*" role="Then">
         <parent parentId="1" distance="1"/>
     </entity>
     <action entityId="3" type="UNCHANGED_HIGH_PRIORITY"/>
