+++ src/main/resources/repairability/Nopol/if_upd_s.xml
@@ -7,7 +7,7 @@
     </entity>
     <action entityId="2" type="UPD"/>
 
-    <entity id="3" type="Block" role="Then">
+    <entity id="3" type="*" role="Then">
         <parent parentId="1" distance="1"/>
     </entity>
     <action entityId="3" type="UNCHANGED_HIGH_PRIORITY"/>
